requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
alembic~=1.4.3